move showLog utils
